+++ CHANGELOG.md
@@ -6,9 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [PEP 440](https://www.python.org/dev/peps/pep-0440/)
 and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.7.1]
+## [0.8.0]
 
 ### Added
+* New `time_series` workflow for time series processing of GSLC stacks.
+
+### Changed
+* The `back_projection` workflow now accepts an optional `--bounds` parameter to specify the DEM extent
 * The back-projection product now includes the elevation.dem.rsc file.
 
 ## [0.7.0]
@@ -79,4 +83,3 @@ and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 * Initial version of repository.
-